+++ packages/react-wrap-with/__tests__/types/children.missing.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import wrapWith from '../../src/wrapWith';
+import { wrapWith } from '../../src/index';
 
 const Header = (_: Record<string, never>) => <h1>Hello, World!</h1>;
