+++ packages/react-wrap-with/__tests__/types/extract.noChildren.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import wrapWith, { ExtractProp, type HowOf } from '../../src/wrapWith';
+import { ExtractProp, type HowOf } from '../../src/index';
 
 import type { ReactNode } from 'react';
 
@@ -9,4 +9,4 @@ const Header = (_: { children?: ReactNode | undefined }) => <h1>Hello, World!</h
 // Initial props must not have "children".
 
 // @ts-expect-error Type 'symbol' is not assignable to type 'undefined'.
-wrapWith(Header, { children: ExtractProp } satisfies HowOf<typeof Header>);
+({ children: ExtractProp }) satisfies HowOf<typeof Header>;
