+++ packages/react-wrap-with/__tests__/types/how.extract.wrongType.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { type HowOf, SpyProp } from '../../src/wrapWith';
+import { type HowOf, SpyProp } from '../../src/index';
 
 import type { PropsWithChildren } from 'react';
