+++ packages/react-wrap-with/__tests__/types/initial.noChildren.tsx
@@ -1,6 +1,6 @@
 import React, { Fragment } from 'react';
 
-import wrapWith, { type HowOf } from '../../src/wrapWith';
+import { type HowOf } from '../../src/index';
 
 import type { ReactNode } from 'react';
 
@@ -9,4 +9,4 @@ const Header = ({ children }: { children?: ReactNode | undefined }) => <h1>{chil
 // "children" cannot be set in how.
 
 // @ts-expect-error Type 'Element' is not assignable to type 'undefined'.
-wrapWith(Header, { children: <Fragment /> } satisfies HowOf<typeof Header>);
+({ children: <Fragment /> }) satisfies HowOf<typeof Header>;
