+++ packages/react-wrap-with/__tests__/types/initial.optional.empty.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import wrapWith, { type HowOf } from '../../src/wrapWith';
+import { type HowOf, wrapWith } from '../../src/index';
 
 import type { PropsWithChildren } from 'react';
