+++ packages/react-wrap-with/__tests__/types/props.wrapped.missing.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import wrapWith from '../../src/wrapWith';
+import { wrapWith } from '../../src/index';
 
 import type { ReactNode } from 'react';
