+++ packages/react-wrap-with/__tests__/types/spy.extraneous.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import wrapWith, { type HowOf, SpyProp } from '../../src/wrapWith';
+import { type HowOf, SpyProp } from '../../src/index';
 
 import type { ReactNode } from 'react';
 
@@ -9,4 +9,4 @@ const Header = ({ children }: { children?: ReactNode | undefined }) => <h1>{chil
 // "xyz" is not a prop of <Header> and cannot be extracted.
 
 // @ts-expect-error Object literal may only specify known properties, and 'className' does not exist in type 'HowOf<({ children }: { children?: ReactNode; }) => Element>'.
-wrapWith(Header, { className: SpyProp } satisfies HowOf<typeof Header>);
+({ className: SpyProp }) satisfies HowOf<typeof Header>;
