+++ packages/react-wrap-with/src/ExtractProp.ts
@@ -1,0 +1,3 @@
+const ExtractProp = Symbol.for('react-wrap-with/ExtractProp');
+
+export default ExtractProp;
