+++ packages/react-wrap-with/src/SpyProp.ts
@@ -1,0 +1,3 @@
+const SpyProp = Symbol.for('react-wrap-with/SpyProp');
+
+export default SpyProp;
