+++ packages/react-wrap-with/src/index.ts
@@ -1,3 +1,7 @@
-import wrapWith from './wrapWith';
+import { HowOf } from './HowOf';
+import ExtractProp from './ExtractProp';
+import SpyProp from './SpyProp';
+import wrapWith from './private/wrapWith';
 
-export { wrapWith };
+export { ExtractProp, SpyProp, wrapWith };
+export { type HowOf };
