+++ packages/react-wrap-with/src/private/type/PropsOf.ts
@@ -1,0 +1,3 @@
+import { type ComponentType } from 'react';
+
+export type PropsOf<T> = T extends ComponentType<infer P> ? P : never;
