+++ packages/react-wrap-with/src/private/type/RefOf.ts
@@ -1,0 +1,3 @@
+import { type RefAttributes } from 'react';
+
+export type RefOf<T> = T extends RefAttributes<infer R> ? R : never;
