+++ packages/react-wrap-with/src/private/util/withEntries.ts
