+++ packages/react-wrap-with/src/private/wrapWith.extract.spec.tsx
@@ -4,7 +4,8 @@
 import { render } from '@testing-library/react';
 import React from 'react';
 
-import wrapWith, { ExtractProp } from './wrapWith';
+import ExtractProp from '../ExtractProp';
+import wrapWith from './wrapWith';
 
 import type { PropsWithChildren } from 'react';
