+++ packages/react-wrap-with/src/private/wrapWith.refCallback.classComponent.spec.tsx
@@ -2,9 +2,10 @@
 /// <reference types="@types/jest" />
 
 import { render } from '@testing-library/react';
-import React, { Component, useEffect, useRef } from 'react';
+import React, { Component } from 'react';
 
-import wrapWith, { ExtractProp } from './wrapWith';
+import ExtractProp from '../ExtractProp';
+import wrapWith from './wrapWith';
 
 import type { PropsWithChildren, Ref } from 'react';
 
@@ -32,26 +33,29 @@ test('ref of RefObject should be passed', () => {
   // GIVEN: Wrapping <Hello> with <Effect effect="blink">.
   const BlinkingHello = wrapWith(Effect, { containerRef: ExtractProp, effect: 'blink' } as const)(Hello);
 
-  const App = ({ onRef }: { onRef: (refs: [HTMLSpanElement | null, Hello | undefined]) => void }) => {
-    const containerRef = useRef<HTMLSpanElement | null>(null);
-    const ref = useRef<Hello>();
-
-    useEffect(() => onRef([containerRef.current, ref.current]), [onRef]);
-
-    return <BlinkingHello containerRef={containerRef} ref={ref} />;
-  };
+  const App = ({
+    onContainerRef,
+    onRef
+  }: {
+    onContainerRef: (ref: HTMLSpanElement) => void;
+    onRef: (ref: HTMLHeadingElement) => void;
+  }) => <BlinkingHello containerRef={onContainerRef} ref={onRef} />;
 
   const handleRef = jest.fn();
+  const handleContainerRef = jest.fn();
 
   // WHEN: Render.
-  render(<App onRef={handleRef} />);
+  render(<App onContainerRef={handleContainerRef} onRef={handleRef} />);
 
-  // THEN: Ref should be passed.
+  // THEN: "containerRef" should be passed.
+  expect(handleContainerRef).toBeCalledTimes(1);
+
+  // THEN: "containerRef" should point to <span>.
+  expect(handleContainerRef.mock.calls[0][0]).toHaveProperty('tagName', 'SPAN');
+
+  // THEN: "ref" should be passed.
   expect(handleRef).toBeCalledTimes(1);
 
-  // THEN: "containerRef" is pointing to the instance of <span>.
-  expect(handleRef.mock.calls[0][0][0]).toHaveProperty('tagName', 'SPAN');
-
-  // THEN: "ref" is pointing to the instance of <Hello>.
-  expect(handleRef.mock.calls[0][0][1]).toBeInstanceOf(Hello);
+  // THEN: "ref" should point to <Hello>.
+  expect(handleRef.mock.calls[0][0]).toBeInstanceOf(Hello);
 });
