+++ packages/react-wrap-with/src/wrapWith.ts
@@ -1,0 +1,3 @@
+import wrapWith from './private/wrapWith';
+
+export default wrapWith;
